pod_containers/text_to_sql/requirements.txt
fastmcp==2.12.5
pandas==2.2.3
sqlalchemy==2.0.41
aiosqlite==0.21.0
uvicorn==0.35.0
requests==2.32.3
fastapi==0.115.14
plotly==6.3.0
ipykernel==7.1.0
msal==1.34.0
python-dotenv
langchain==0.3.26 \
langchain-core==0.3.81 \
langchain-openai==0.3.28 \
langchain-groq==0.3.0 \
langchain-mcp-adapters==0.1.9 \
langgraph==0.6.7 \
openai<2.0